The class Arrays includes a lot of useful methods for processing arrays. One of them is the binarySearch method. It allows you to quickly find the index of an element in the given ordered array. You just need to pass the array of values and the value to search for to this method.
In this quiz, you need to call this method from the Arrays class without using an import statement.

Do not use your own implementation of the searching method, it may not work with prepared tests.

Clock clock = new Clock(); // the time is 12:00
clock.next(); // the time is 12:01
2) The initial time of a clock object is 10:59, i.e. hours = 10 and minutes = 59. After applying the next method to the clock, we turn its fields to hours = 11 and minutes = 0, i.e., the time is 11:00 now.

Clock clock = new Clock();
clock.hours = 10;
clock.minutes = 59;
// time is 10:59

clock.next(); // the time is 11:00