Given a method that takes a long value and returns the max value of the long type minus the given number.
public static long getMaxMinusCurrent(long val) {
    return Long.MAX_VALUE - val;
}
Overload this method with one new method. The method should take an int and return the max value of the int type minus the given value.

For the method, the type of result value must be the same as the type of the parameter.