When the AI is playing at medium difficulty level, it makes moves using the following logic:
If it already has two in a row and can win with one further move, it does so.
If its opponent can win with one move, it plays the move necessary to block this.
Otherwise, it makes a random move.
You should add a medium parameter so that you can play against this level. It should also be possible to make AIs using easy and medium levels play against each other!